+++ src/main/java/net/minecraft/world/NextTickListEntry.java
@@ -12,7 +12,7 @@ public class NextTickListEntry implements Comparable<NextTickListEntry> {
     private final long tickEntryID;
 
     public NextTickListEntry(BlockPos positionIn, Block blockIn) {
-        this.tickEntryID = (long) (nextTickEntryID++);
+        this.tickEntryID = nextTickEntryID++;
         this.position = positionIn;
         this.block = blockIn;
     }
